requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1

alembic==1.14.0

pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

python-jose==3.3.0
python-multipart==0.0.20

passlib==1.7.4
argon2-cffi==23.1.0

pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
